+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support for globally defined translations. Page item keys can incllude the namespace `global:itemKey`, so that the translation is not searched on the current page's json file, but in the `global.json`.
+- New page item type `actionCard` that utilises the ActionCard component from case-web-ui.
 
 ### Changed